Feat:Todos relation with users

+++ src/app/api/todos/route.ts
@@ -1,4 +1,5 @@
 
+import { getUserSessionServer } from '@/auth/actions/auth-actions'
 import prisma from '@/lib/prisma'
 import {NextResponse, NextRequest} from 'next/server'
 
@@ -37,10 +38,16 @@ const postSchema = yup.object({
 
 export async function POST(request: Request){
 
+  const user = await getUserSessionServer()
+
+  if(!user){
+    return NextResponse.json('No autorizado',{ status: 401 })
+  }
+
   try{
     const {complete, description} = await postSchema.validate( await request.json())
 
-    const todo = await prisma.todo.create({ data: {complete, description} })
+    const todo = await prisma.todo.create({ data: {complete, description, userID: user.id} })
   
   
     return NextResponse.json(todo)
@@ -53,10 +60,16 @@ export async function POST(request: Request){
 }
 
 export async function DELETE(request: Request){
+  const user = await getUserSessionServer()
+
+  if(!user){
+    return NextResponse.json('No autorizado',{ status: 401 })
+  }
+
 
   try{
 
-     await prisma.todo.deleteMany({ where: {complete: true} })
+     await prisma.todo.deleteMany({ where: {complete: true, userID: user.id} })
   
   
     return NextResponse.json('Borrados')
